prac1/app/requirements.txt
avro
flask
jsonpickle
msgpack
protobuf
pympler
xml_marshaller
pyyaml
